scraper/pyproject.toml
[project]
name = "jobportal-scraper"
version = "0.1.0"
description = "JobSpy + Bundesagentur scraper with Supabase upsert"
requires-python = ">=3.12"
dependencies = [
    "python-jobspy>=1.1.79",
    "supabase>=2.10.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.7.0",
]

[project.scripts]
scrape = "scraper.run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scraper"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
